worked on JDBC, JPA S3Buckets, UserService and User Controller

+++ backend/src/main/java/com/rolez/backend/users/UserDao.java
@@ -5,12 +5,12 @@ import java.util.Optional;
 
 public interface UserDao {
     List<User> selectAll();
-    Optional<User> selectUserById(Long id);
+    Optional<User> selectUserById(Integer id);
     void insertUser(User user);
     boolean existsUserByEmail(String email);
-    boolean existsUserById(Long id);
-    void deleteUserById(Long id);
+    boolean existsUserById(Integer id);
+    void deleteUserById(Integer id);
     void updateUser(User update);
     Optional<User> selectUserByEmail(String email);
-    void updateUserProfileImageId(String profileImageUrl, Long id);
+    void updateUserProfileImageId(String profileImageUrl, Integer id);
 }
